fix n/a on admin schedules

+++ database/migrations/2025_02_10_085840_add_preferred_time_to_schedules_table.php
@@ -4,23 +4,31 @@ use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
 
-return new class extends Migration
-{
+return new class extends Migration {
     /**
      * Run the migrations.
      */
     public function up()
-{
-    Schema::table('schedules', function (Blueprint $table) {
-    
-        $table->string('file_name')->nullable(); 
-    });
-}
+    {
+        Schema::table('schedules', function (Blueprint $table) {
+            if (!Schema::hasColumn('schedules', 'file_name')) {
+                $table->string('file_name')->nullable();
+            }
+            if (!Schema::hasColumn('schedules', 'preferred_time')) {
+                $table->enum('preferred_time', ['AM', 'PM'])->default('AM');
+            }
+        });
+    }
 
-public function down()
-{
-    Schema::table('schedules', function (Blueprint $table) {
-        $table->dropColumn('preferred_time');
-    });
-}
+    /**
+     * Reverse the migrations.
+     */
+    public function down()
+    {
+        Schema::table('schedules', function (Blueprint $table) {
+            if (Schema::hasColumn('schedules', 'preferred_time')) {
+                $table->dropColumn('preferred_time');
+            }
+        });
+    }
 };
